Pulled fixtures back into module
